Bug 1940353 - Fix verifyRedesignedMenuAfterDisablingAnExtensionTest UI test r=ohorvath
Regression caused by [[ https://bugzilla.mozilla.org/show_bug.cgi?id=1925090 | 1925090 ]]

The UI test was failing because the "Successfully disabled addonName" snackbar was covering the nav bar when trying to open and expand the redesigned main menu. (no position change, the snackbar seems to be displayed for a longer period of time)

Added a wait for the snack bar to be gone right after disabling the add-on.

The UI test successfully passed 25x on Firebase ✅

Differential Revision: https://phabricator.services.mozilla.com/D233477