Bug 1948004 - Fix verifyTheShareButtonTest UI test r=aaronmt
The UI test failed when trying to verify the sharing tab intent.
This was happening because the page wasn't fully loaded before actually sharing it.

To overcome this problem, I've added a page content check to be sure the page was fully loaded.

The UI test successfully passed 100x on Firebase ✅

Differential Revision: https://phabricator.services.mozilla.com/D238197