Bug 1948032: Enable data-deletion request before sending it  a=dmeehan
The deletion request needs to be enabled after the usage reporting is enabled, because that will set up the usage reporting id which will identify the deletion request.
There's no reason for the deletion request to ever be disabled, however, so it's just set to `true`.

Original Revision: https://phabricator.services.mozilla.com/D238070

Differential Revision: https://phabricator.services.mozilla.com/D238173