Bug 1949173 - Re-enable UI tests affected by network connection problems r=aaronmt,ohorvath
We've disabled a couple of UI tests in which we enable/disable the network connection because it caused problems with other running tests.

This happened because after there's a delay in enabling/disabling the network connection using the commands.

To overcome this problem, I've created a timer idling resource which ensures that the test framework waits for a specified amount of time before proceeding with the next steps.

After the TimerIdlingResource becomes idle, we'll check the active network state (Determine the type of active network connection and check that the active network is connected to WiFi or Cellular) using a newly created checkActiveNetworkState function

All 11 disabled UI tests successfully passed 200x on Firebase ✅
All UI tests successfully passed 1x on Firebase ✅

Differential Revision: https://phabricator.services.mozilla.com/D238783