Bug 1931556 - Part 2: Use primary button color in extension prompts r=android-reviewers,willdurand,Roger
In preparation for adding a "disabled" button style, I noticed that the
current button colors follow an old design system. This patch updates
the colors to the latest version, and defines the aliases for the
"disabled" style, which will be used in the next patch.

The current button in extension permission prompts background has the
"accent" color because it was the design at the time of introduction, in
https://github.com/mozilla-mobile/fenix/issues/8593

The "action primary" style was introduced in
https://github.com/mozilla-mobile/fenix/issues/24055
... but not applied to the existing button.

Comparison of colors (for regular and private browsing):

```
Before: R.attr.accent
- @color/accent_normal_theme  = @color/photonViolet50
- @color/accent_private_theme = @color/photonViolet50
After: R.attr.actionPrimary
- @color/fx_mobile_action_color_primary         = @color/photonInk20
- @color/fx_mobile_private_action_color_primary = @color/photonViolet60
New: R.attr.actionPrimaryDisabled
- @color/fx_mobile_action_color_primary_disabled         = @color/photonInk20A20
- @color/fx_mobile_private_action_color_primary_disabled = @color/photonViolet60A50 (from fx_mobile_action_color_primary_disabled in values-night/colors.xml)

Before: R.attr.textOnColorPrimary
- @color/fx_mobile_text_color_oncolor_primary         = @color/photonLightGrey05
- @color/fx_mobile_private_text_color_oncolor_primary = @color/photonLightGrey05
After: R.attr.textActionPrimary
- @color/fx_mobile_text_color_action_primary         = @color/photonLightGrey05
- @color/fx_mobile_private_text_color_action_primary = @color/photonLightGrey05
New: R.attr.textActionPrimaryDisabled
- @color/fx_mobile_text_color_action_primary_disabled = @color/photonDarkGrey90A50
- @color/fx_mobile_private_text_color_action_primary_disabled = @color/photonLightGrey05A40 (from fx_mobile_text_color_action_primary_disabled in values-night/colors.xml)
```

Differential Revision: https://phabricator.services.mozilla.com/D238306