Bug 1825296 - Fix customizedTrackingProtectionOptionsTest UI test r=ohorvath
Both **customizedTrackingProtectionOptionsTest** and **verifyTrackersBlockedWithCustomTPOptionsDisabledTest** were flaky over the weekend because the "Known Fingerprinters" check box couldn't be properly verified that it's displayed.

To fix the problem I've added a scroll action to make sure the options and check boxes from the lower part of the view are visible.

Both UI tests successfully passed 100x on Firebase ✅
The other 3 UI tests that are using the **verifyCustomTrackingProtectionSettings** function successfully passed 25x on Firebase ✅

Differential Revision: https://phabricator.services.mozilla.com/D245406