Bug 1960122 - Use `MOZ_BUILD_DATE` in Fenix build configuration. r=adhingra,tcampbell,aaronmt,android-reviewers
This commit uses `MOZ_BUILD_DATE`.  `MOZ_BUILD_DATE` is 14 characters,
"YYYYMMDDHHMMSS".  That is not the same as the old ISO 8601 format.
But there's no need to use ISO 8601; any representation is fine, and
this is the Mozilla standard presentation.

While here, remove a test that's not actually running and of
questionable value.

This duplicates `getBuildId()` yet again.  Eventually, we'll be able
to unify all of these `getBuildId()` duplications into a single
Mozilla-wide plugin, but that will be much easier after we have a
single top-level `settings.gradle`.

Differential Revision: https://phabricator.services.mozilla.com/D247502