Bug 1958676 - avoid clobbering user-defined values of post-quantum key exchange prefs on android. r=keeler,geckoview-reviewers,android-reviewers,ohall
Previously the security.tls.enable_kyber and network.http.http3.enable_kyber
prefs would be updated at startup to the default value of
postQuantumKeyExchangeEnabled (defined in fenix/app/nimbus.fml.yaml). This
patch makes postQuantumKeyExchangeEnabled a tri-state with a default value that
indicates that the prefs should not be modified. The other two states can still
be used to control the prefs as part of a Nimbus experiment.

Differential Revision: https://phabricator.services.mozilla.com/D247375