Bug 1961377 - Adding POM Scaffolding r=ajoltan
This is the initial scaffolding for setting up the use of a Page Object Model design.

These intial helper modules prepare the creation of Test Factories leveraging data-driven testing patterns and test runtime configuration for dynamic test creation and execution.

This prepares the way for targeted test suites on an ad-hoc basis per the needs of the state and context-under-test.

Differential Revision: https://phabricator.services.mozilla.com/D246044