Bug 1968912 - Fix verifyMainMenuItemsTest efficiency UI test r=jajohnson,aaronmt,ohorvath
The UI test started to fail after [[ https://phabricator.services.mozilla.com/D249266 | D249266 ]] landed.

Because the "New private tab" button is not displayed anymore in the composable main menu I've changed it with the bookmarks button.

All UI tests successfully passed on the experimental task. ✅

I've also update the description of the button as it was incorrect.

Differential Revision: https://phabricator.services.mozilla.com/D251499