Bug 1855893 - Fix installAddonFromMainMenuTest UI test r=aaronmt
The UI test was flaky when trying to verify the "Downloading add-on" prompt was displayed.
This happened because the prompt in some cases wasn't displayed or disappeared very fast, most likely because uBlock was downloaded right away.

To overcome this problem, I've swapped out uBlock with AdGuard AdBlocker.
When downloading AdGuard AdBlocker takes a bit more time and we are able to verify the prompt.

The UI test successfully passed 100x on Firebase ✅

Differential Revision: https://phabricator.services.mozilla.com/D252315