Bug 1971166 - part 1 - Highlight the domain of the current displayed URL r=android-reviewers,pollymce
Similar to the toolbar View functionality the domain of the current
displayed URL will be shown with a solid color while the rest of the
URL will be shown with a faded color.
The displayed URL will also be scrolled to ensure the domain is always
visible even if the URL has longer subdomains or pages and the ends
will be shown as faded.

Differential Revision: https://phabricator.services.mozilla.com/D254939