Bug 1976535 - Use the same BrowserToolbarStore builder for both the navbar and the addressbar r=android-reviewers,Roger
Previously both the navbar and the addressbar would build their own BrowserToolbarStore
but have this persisted as a ViewModel in the containing fragment.
This meant that whoever would try to build the store the second time would get the
initially built store, with potentially different middlewares set.

This patch brings one clear API to build the BrowserToolbarStore that can be used by multiple
separate features which would get the store instance properly configured to support them all.

Differential Revision: https://phabricator.services.mozilla.com/D257694