Bug 1872421 - Add Nimbus experiments to crash report runtime tags r=android-reviewers,boek
This change introduces a new `NimbusExperimentsRuntimeTagProvider` that adds active Nimbus experiments as runtime tags to crash reports.

This will help in analyzing crashes by providing context about which experiments were active at the time of the crash.

The `NimbusExperimentsRuntimeTagProvider` is added to the list of `runtimeTagProviders` in the `Analytics` component.

Differential Revision: https://phabricator.services.mozilla.com/D258411