Bug 1977303 - RequestWindowFeature for SearchDialogFragment r=calu,android-reviewers
We need to request the window feature that we are using while also
using the `DialogFragment.STYLE_NORMAL` because after bug 1976746,
we are no longer avoiding insets being applied to lower API
versions.

This is fine without tests since we are going to raise our minSDK
and the only reasonable way I can think of writing a test is to using
this SDK level in UI tests which would make the rest of the tests that
run on this level flaky.

I've tested this locally against an API 23 and 36 emulator; will request
QA testing as well when it lands.

Differential Revision: https://phabricator.services.mozilla.com/D258528