Bug 1977818 - Fix verifyTheExtensionsdMenuListAfterRemovingAnExtensionTest UI test  a=diannaS
Both **verifyTheExtensionsdMenuListAfterRemovingAnExtensionTest** and **verifyTheManageExtensionsItemTest** were flaky when trying to :

  - Verify that the "Bitwarden" extension is installed
  - Open the details of the "Bitwarden" extension after being installed

This was happening because of the inconsistencies of the extensions naming across the app.
To fix this, I've added a check that will update the extension name to use the entire title in the above mentioned sections of the app.

There were also some sync issues when trying to open the "Extension" sub menu from the main menu (the list of recommended extensions wasn't always displayed)

Both UI tests successfully passed 100x on Firebase ✅

Original Revision: https://phabricator.services.mozilla.com/D258551

Differential Revision: https://phabricator.services.mozilla.com/D258664