Bug 1977755 - Fix verifyTheManageExtensionsItemTest UI test r=aaronmt
The UI test failed when trying to open again the generic page and when trying to verify the page content after opening the "Manage extensions" section.

This was happening because for some extensions a new tab is being opened with the website of that extension.
  - the extensions website was simultaneously opened with the new generic page
  - when trying to verify the page content the extensions website was actually being displayed for some extensions instead of the generic page

To fix these problems I've:
  - Removed the redundant page loading of the generic page the second time
  - Removed the redundant page verification at the end of the test

The UI test successfully passed 100x on Firebase ✅

Differential Revision: https://phabricator.services.mozilla.com/D259091