Bug 1979823 - Fix setting initial state of PrivateBrowsingLockFeature r=twhite,android-reviewers
Now that we aren't using the application instance of the feature to hold
our state, we should read the AppStore value during init similar to our
persisted feature enabled flag.

We don't really need to check whether there are private tabs AND
`isPrivateScreenLocked` because private tabs are never persisted, but I
left that original logic in there because it doesn't seem to hurt to
have it.

Differential Revision: https://phabricator.services.mozilla.com/D259205