Bug 1948756 - part 1 - Use just one toolbar environment r=android-reviewers,Roger
We use the Environment idiom to be able to pass lifecycle based dependencies
to middlewares.
Since the browser toolbar is currently used in different contexts and with
different functionalities in the home and browser screens the concrete
environment classes were different.
But with searches being possible for search terms in the browser screen also
the search middlewares need to work with only one environment type, which now
will be BrowserToolbarEnvironment.

Differential Revision: https://phabricator.services.mozilla.com/D259418