Bug 1984736 - Part 4: Add support for subtitle to app icon option r=android-reviewers,rebecatudor273
The Customise screen grants the user the ability to change the app icon, and those options can have just a title for its name, but some might have a subtitle as well, for additional information about the icon option. Adjusted the app icon option composable element to draw both variants.

Differential Revision: https://phabricator.services.mozilla.com/D263030