Bug 1988105 - part 2 - Simpler way of padding the home content r=android-reviewers,devota
All widgets on the homescreen are placed absolutely on the same canvas.
This means that the home content would be clipped by the top/bottom
toolbar or other such widgets and to prevent that it needs to have the
appropriate vertical margins set.
We can do this in just one place for all versions of the home content
for simpler yet more powerful functionality by also supporting the
different screen orientations for which the content needs different
margins as other widgets may be shown for each orientation.
One scenario which is not supported it that of the microsurvey
being shown/closed, this is to be fixed in bug 1988142.

Differential Revision: https://phabricator.services.mozilla.com/D264567