Bug 1977632 - Fix verifyRecommendedExtensionsListWhileNoExtensionIsInstalledTest UI test  a=RyanVM DONTBUILD
The UI test was failing because the "Extensions" manager was opened when clicking the "Extensions" button from the composable main menu.

To fix the problem I've added a couple of assertions to improve the sync between actions.

The UI test successfully passed 50x on Firebase ✅

Original Revision: https://phabricator.services.mozilla.com/D266992

Differential Revision: https://phabricator.services.mozilla.com/D268119