Bug 1999589 - Fix verifyClockFormInteractionsTest UI test r=mcarare
The UI test was failing on Firebase because the time was sometimes set by default to PM, meaning that when we set the time to 10:10 it would actually set it to 22:10 and we were checking for 10:10
To fix this problem, I've added an extra step, namely to always set the time to AM.

Another problem noticed while checking locally, is that the time couldn't actually be selected properly due to some selector changes. (not sure why this hasn't occurred when Gabi did the first changes)
To fix this I've updated the selectors used to select the time.

The UI test [[ https://treeherder.mozilla.org/jobs?repo=try&selectedTaskRun=DL5hCDyxTeGVmr-6TCirWg.0&revision=2201050a90d520d1ccc37a50c8928926727435f6 | successfully passed ]] 25x on Firebase ✅

Differential Revision: https://phabricator.services.mozilla.com/D272217