Bug 1995706 - Fix verifyStateFieldUpdatesInAccordanceWithCountryFieldTest UI test r=aaronmt
The UI test was flaky because it couldn't click properly with the "country" dropdown because the keyboard remained displayed.
Added a couple of 1s waits for the app window to be updated for better syncing after opening the "Add address" section and after dismissing the keyboard.

The UI test successfully passed 100x on Firebase ✅

Differential Revision: https://phabricator.services.mozilla.com/D274280